Fixed wrong processor assignment
 - Added Scores on Ui
 - Score modifications

+++ include/Elpida/Engine/Result/BenchmarkResult.hpp
@@ -37,7 +37,7 @@ namespace Elpida
 	class BenchmarkResult
 	{
 	public:
-		using Score = size_t;
+		using Score = double;
 
 		[[nodiscard]] const Benchmark& getBenchmark() const
 		{
